moved Bootstrapper to __init__.py
